Feat: refactor with breakpoints
